natifylambda/requirements.txt
aws-cdk-lib>=2.99.0
aws-cdk.aws-lambda-python-alpha>=2.99.0-alpha.0
constructs>=10.0.0
